Mavenize and update to 1.7.10
