--- config.yml
@@ -1,30 +1,0 @@
-#Builder configuration
-
-#The directory to look for schematics. 
-SchematicsFolder: plugins/Builder/schematics/
-
-#The allowed materials for plot marking, the first one in the list is the default. 
-#Use Id or Name. Should only use base blocks, not things that can break on place (like TORCH).
-MarkMaterials: 
- - GLASS
- - GLOWSTONE
- - BEDROCK
-  
-#Customize the messages Builder sends. <NPC>, <SCHEMATIC>, <AMOUNT>, <PLAYER>, and <ITEM> can be used as placeholders where appropriate.
-DefaultTexts:
- BuildStarted: "&a<NPC> has started to build &f<SCHEMATIC>!"
- BuildComplete: "&a<NPC> has completed build of &f<SCHEMATIC>!"
- BuildCanceled: "&c<NPC>'s build of &f<SCHEMATIC>&c has been canceled!"
- BuildCollecting: "&6<NPC> needs &f<AMOUNT> &6more blocks to start building &f<SCHEMATIC>."
- Mark: "&a<NPC> is marking the corners of &f<SCHEMATIC>"
- Survey: "&a<NPC> needs these materials to build &f<SCHEMATIC> &aat this location."
- Supply_List: "&f<NPC> says: &6I still need the following items to build &f<SCHEMATIC>, &6<PLAYER>:"
- Supply_Need_Item: "&f<NPC> says: &aI still need &f<AMOUNT> <ITEM> &afor &f<SCHEMATIC>&9. Right-click again to donate."
- Supply_Dont_Need_Item: "&f<NPC> says: &6I don't need any &f<ITEM>&6. Thanks anyway."
- Supply_Item_Taken: "&f<NPC> says: &aThanks for the &f<AMOUNT> <ITEM>, <PLAYER>!"
-
-DefaultOptions:
-  MoveTimeoutSeconds: 2.0
-  IgnoreProtection: false #not working
-  HoldItems: true
-  RequireMaterials: false
