--- supply.txt
@@ -1,30 +1,0 @@
-Original Block Number:Require Item Number:Amount(Optional)
-Set to 0 to never require block. Blocks not specified will require their normal drop item.
-90:0
-97:0
-95:54
-78:0
-34:0
-119:0
-80:80
-92:354
-43:44:2
-125:126:2
-20:20
-102:102
-47:47
-103:362
-130:130
-134:53
-135:53
-136:53
-128:67
-109:67
-108:67
-79:332
-51:0
-59:295
-72:70
-71:330:0.5
-26:335:0.5
-137:0
